web_regression_test/requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1
urllib3==2.0.7
python-dotenv==1.0.0
PyYAML==6.0.1
pandas==2.1.3

error_detection:
  ignore_patterns:
    - "favicon.ico"
    - "net::ERR_FAILED"
  screenshot_on_error: true
  error_keywords:
    - "error"
    - "异常"
    - "失败" 